requirements.txt
streamlit
openai
langchain
beautifulsoup4
playwright
tldextract
pandas
certifi
python-dotenv
langchain_community